pyproject.toml
[tool.poetry]
name = "cargpt"
version = "0.2.0"
description = ""
authors = ["Evgenii Gorchakov <evgenii@yaak.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
torch = "^2.2.0"
torchvision = "^0.17.0"
torchaudio = "^2.2.0"
tensordict = "^0.3.0"
xformers = "=0.0.24"
pytorch-lightning = "^2.1.4"
einops = "^0.7.0"
hydra-core = "^1.3.0"
loguru = "^0.6.0"
jaxtyping = "^0.2.23"
beartype = "*"
more-itertools = "^9.0.0"
torchinfo = "^1.7.2"
wandb = "*"
pillow = ">=9.5.0"
dill = "^0.3.6"
tqdm = "=4.65.0"
transformers = "=4.30.2"

# lightning problems
scipy = "^1.11.4"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.train]
optional = true

[tool.poetry.group.train.dependencies]
omegaconf = "^2.3.0"
yaak-datasets = {git = "git@github.com:yaak-ai/yaak-datasets.git", rev = "v0.17.1"}
polars = "*"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
pyright = "*"
ruff = "*"
deptry = "*"
opencv-stubs = "*"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ipdb = "*"
ipython = "*"
pudb = "*"
ipython-autoimport = "*"
pyfzf = "*"
ruff-lsp = "*"
pyright = "*"
lovely-tensors = "*"
pre-commit = "*"

[tool.poetry.group.notebook]
optional = true

[tool.poetry.group.notebook.dependencies]
jupyterlab = "^4.1.1"
jupyterlab-vim = "^4.1.3"
hvplot = "^0.9.2"

[tool.poetry.group.inference]
optional = true

[tool.poetry.group.inference.dependencies]
rerun-sdk = "*"
opencv-python = "*"

[tool.poetry.group.dalle]
optional = true

[tool.poetry.group.dalle.dependencies]
dall-e = {git = "git@github.com:openai/DALL-E.git", rev = "master"}

[tool.poetry.group.log]
optional = true

[tool.poetry.group.log.dependencies]
tensorboard = "^2.16.0"

[tool.pyright]
typeCheckingMode = "basic"
reportUnusedImport = true
reportUnusedVariable = true
reportUnusedFunction = true
reportMissingSuperCall = true
reportUnnecessaryIsInstance = true
reportUnnecessaryCast = true
reportUnnecessaryTypeIgnoreComment = true
reportUnnecessaryComparison= true
reportUnusedCallResult = true
reportUnusedExpression = true
reportMatchNotExhaustive = true

exclude = ['wandb/', 'cargpt/models/gato.py']

[tool.ruff]
preview = true
unsafe-fixes = true

[tool.ruff.lint]
select = ["ALL"]
fixable = ["ALL"]
ignore = ["E501", "F722", "F821", "E203", "D", "CPY", "ANN", "PGH", "FA", "PLR", "N", "TD", "FIX", "S101", "SLF", "C901", "COM812", "ISC001", "A001", "A002", "PD", "FBT003", "EXE", "SIM118", "NPY002"]

[tool.ruff.lint.isort]
known-third-party = ["wandb"]


[tool.deptry.per_rule_ignores]
DEP001 = ["rerun", "deephouse", "fiftyone"]
DEP002 = ["dill", "scipy", "torchaudio"]
DEP003 = ["numpy", "typing_extensions", "numpy", "lightning_utilities"]
DEP004 = ["cv2", "dall_e", "omegaconf", "yaak_datasets", "polars"]
